Fix e2e screenshot on timeout failure
